# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: E-commerce platform API with WhatsApp AI integration
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
        "401":
          description: Unauthorized

  # Products
...
  /customers:
    get:
      operationId: listCustomers
      tags: [customers]
      summary: List customers
      parameters:
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of customers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Customer"
    post:
      operationId: createCustomer
      tags: [customers]
      summary: Create or upsert customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCustomerBody"
      responses:
        "201":
          description: Customer created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
  /customers/{id}:
    get:
      operationId: getCustomer
      tags: [customers]
      summary: Get a customer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Customer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
        "404":
          description: Not found
    patch:
      operationId: updateCustomer
      tags: [customers]
      summary: Update a customer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCustomerBody"
      responses:
        "200":
          description: Updated customer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"

  # Bookings
  /bookings:
    get:
      operationId: listBookings
      tags: [bookings]
      summary: List bookings
      parameters:
        - name: date
          in: query
          schema:
            type: string
        - name: productId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of bookings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Booking"
    post:
      operationId: createBooking
      tags: [bookings]
      summary: Create a booking
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBookingBody"
      responses:
        "201":
          description: Booking created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
  /bookings/{id}:
    patch:
      operationId: updateBooking
      tags: [bookings]
      summary: Update booking status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBookingBody"
      responses:
        "200":
          description: Updated booking
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"

  # WhatsApp
...
components:
  schemas:
...
    LoginBody:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required: [username, password]
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/AdminUser"
        token:
          type: string
      required: [user, token]
    AdminUser:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        createdAt:
          type: string
      required: [id, username, createdAt]
...
    Customer:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: string
        email:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        totalOrders:
          type: integer
        totalSpent:
          type: number
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, name, phone, totalOrders, totalSpent, createdAt, updatedAt]
    CreateCustomerBody:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        email:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
      required: [name, phone]
    UpdateCustomerBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
    Booking:
      type: object
      properties:
        id:
          type: integer
        orderId:
          type: ["integer", "null"]
        productId:
          type: integer
        productName:
          type: string
        customerId:
          type: ["integer", "null"]
        customerName:
          type: string
        customerPhone:
          type: string
        slotId:
          type: integer
        date:
          type: string
        startTime:
          type: string
        endTime:
          type: string
        status:
          type: string
          enum: [pending, confirmed, cancelled, completed]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, productId, productName, customerName, customerPhone, slotId, date, startTime, endTime, status, createdAt]
    CreateBookingBody:
      type: object
      properties:
        productId:
          type: integer
        slotId:
          type: integer
        customerName:
          type: string
        customerPhone:
          type: string
        notes:
          type: ["string", "null"]
      required: [productId, slotId, customerName, customerPhone]
    UpdateBookingBody:
      type: object
      properties:
        status:
          type: string
          enum: [pending, confirmed, cancelled, completed]
...